+++ sdk/src/main/java/ly/count/android/sdk/ConfigurationProvider.java
@@ -29,4 +29,6 @@ interface ConfigurationProvider {
     int getBOMRequestAge();
 
     int getBOMDuration();
+
+    int getRequestTimeoutDurationMillis();
 }