+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 25.4.4
 * Improved disk size calculation in crash reports.
+* Added "setRequestTimeoutDuration(requestTimeoutDuration)" init config method configure request timeout in seconds.
 
 * Mitigated an issue displaying Content on API level 35 and above.